add workflow, commented headers and seperated long files into small ones
