+++ .github/pull_request_template.md
@@ -1,0 +1,14 @@
+## Assigned Issue Number ? { if does not exist then create one }
+#007
+
+## Language Opted ?
+C/C++/Python/Java/...
+
+## Number of files added or updated ?
+1/2/3/4/...
+
+### Checklist
+
+- [ ] I've read the contribution guidelines.
+- [ ] I've referred the correct issue number.
+- [ ] I've fill up this entire template correctly.
