+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "Display/Display.h"
+#include "Include/Display.h"
 
 int main() {
     Display display;
